requirements.txt
simple-lama-inpainting
gradio
pillow
torch